requirements.txt
chromadb==0.4.22
GitPython==3.1.41
langchain==0.1.0
langchain-openai==0.0.2.post1
lorem==0.1.1
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
sentence-transformers==2.2.2